Revert "CRUD UNIT TESTS AND JSDOCS ARE COMPLETE 😈 "
