+++ .github/workflows/main.yml
@@ -18,6 +18,6 @@ jobs: # the things being executed
       - name: Unit Test
         run: node --experimental-vm-modules node_modules/jest/bin/jest.js admin/cipipeline/unit.test.js
       - name: CRUD Unit Tests
-        run: npm test -- ./admin/cipipeline/CRUDUnit.test.js
+        run: npm test -- ./admin/cipipeline/createCardUnit.test.js
       - name: Reset Cards Unit Test
         run: npm test -- ./admin/cipipeline/resetCards.test.js
